requirements.txt
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
ffmpeg==1.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
PySocks==1.7.1
python-dotenv==1.1.1
python-telegram-bot==21.4
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
urllib3==2.5.0
yt-dlp==2025.9.5
